requirements.txt
attrs==19.3.0
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
Django==2.2.7
django-extensions==2.2.5
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.2
mysqlclient==1.4.6
parsel==1.5.2
pkg-resources==0.0.0
Protego==0.1.15
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
python-scrapyd-api==2.1.2
pytz==2019.3
queuelib==1.5.0
requests==2.22.0
Scrapy==1.8.0
scrapyd==1.2.1
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
Twisted==19.10.0
urllib3==1.25.7
w3lib==1.21.0
zope.interface==4.7.1
